[quant][graphmode] API - add more passes to graph mode
Summary:
Add `insert_prepack_unpack` and `fold_prepack` to `convert_script`

Test Plan:
.

Reviewers:
pt1quant

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
